docs(atomit): move folded selectors to separate file
https://coveord.atlassian.net/browse/KIT-1521
